move Authentication to shared
