+++ androidApp/src/main/java/dev/johnoreilly/confetti/di/AppModule.kt
@@ -15,7 +15,7 @@ import dev.johnoreilly.confetti.SpeakerDetailsViewModel
 import dev.johnoreilly.confetti.SpeakersViewModel
 import dev.johnoreilly.confetti.TokenProvider
 import dev.johnoreilly.confetti.account.AccountViewModel
-import dev.johnoreilly.confetti.account.Authentication
+import dev.johnoreilly.confetti.auth.Authentication
 import dev.johnoreilly.confetti.SearchViewModel
 import dev.johnoreilly.confetti.settings.SettingsViewModel
 import dev.johnoreilly.confetti.wear.WearSettingsSync
