+++ shared/build.gradle.kts
@@ -71,6 +71,7 @@ kotlin {
         val mobileMain by getting {
             dependencies {
                 implementation(libs.kmm.viewmodel)
+                implementation(libs.firebase.mpp.auth)
             }
         }
